Move types directory to root
